+++ modules/sdk/gradle-plugins-target-platform/src/main/java/com/liferay/gradle/plugins/target/platform/TargetPlatformIDEPlugin.java
@@ -24,7 +24,9 @@ import io.spring.gradle.dependencymanagement.dsl.DependencyManagementConfigurer;
 import io.spring.gradle.dependencymanagement.dsl.DependencyManagementExtension;
 import io.spring.gradle.dependencymanagement.dsl.ImportsHandler;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.Map;
@@ -33,16 +35,24 @@ import java.util.Set;
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
+import org.gradle.api.UnknownDomainObjectException;
 import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.artifacts.Dependency;
 import org.gradle.api.artifacts.DependencySet;
-import org.gradle.api.plugins.JavaBasePlugin;
+import org.gradle.api.file.FileCollection;
+import org.gradle.api.plugins.JavaPlugin;
+import org.gradle.api.tasks.SourceSet;
 import org.gradle.plugins.ide.eclipse.EclipsePlugin;
 import org.gradle.plugins.ide.eclipse.model.EclipseClasspath;
 import org.gradle.plugins.ide.eclipse.model.EclipseModel;
+import org.gradle.plugins.ide.idea.IdeaPlugin;
+import org.gradle.plugins.ide.idea.model.IdeaModel;
+import org.gradle.plugins.ide.idea.model.IdeaModule;
+import org.gradle.plugins.ide.idea.model.internal.GeneratedIdeaScope;
 
 /**
  * @author Gregory Amerson
+ * @author Simon Jiang
  */
 public class TargetPlatformIDEPlugin implements Plugin<Project> {
 
@@ -54,7 +64,8 @@ public class TargetPlatformIDEPlugin implements Plugin<Project> {
 	@Override
 	public void apply(Project project) {
 		GradleUtil.applyPlugin(project, EclipsePlugin.class);
-		GradleUtil.applyPlugin(project, JavaBasePlugin.class);
+		GradleUtil.applyPlugin(project, JavaPlugin.class);
+		GradleUtil.applyPlugin(project, IdeaPlugin.class);
 		GradleUtil.applyPlugin(project, TargetPlatformPlugin.class);
 
 		DependencyManagementExtension dependencyManagementExtension =
@@ -71,6 +82,7 @@ public class TargetPlatformIDEPlugin implements Plugin<Project> {
 				targetPlatformIDEExtension);
 
 		_configureEclipseModel(project, targetPlatformIDEConfiguration);
+		_configureIdeaModel(project, targetPlatformIDEConfiguration);
 	}
 
 	private Configuration _addConfigurationTargetPlatformIDE(
@@ -217,4 +229,50 @@ public class TargetPlatformIDEPlugin implements Plugin<Project> {
 		plusConfigurations.add(targetPlatformIDEConfiguration);
 	}
 
+	private void _configureIdeaModel(
+		Project project, Configuration targetPlatformIDEConfiguration) {
+
+		IdeaModel ideaModel = GradleUtil.getExtension(project, IdeaModel.class);
+
+		IdeaModule ideaModule = ideaModel.getModule();
+
+		Map<String, Map<String, Collection<Configuration>>> scopes =
+			ideaModule.getScopes();
+
+		Map<String, Collection<Configuration>> providedScope = scopes.get(
+			GeneratedIdeaScope.PROVIDED.name());
+
+		if (providedScope == null) {
+			providedScope = new HashMap<>();
+		}
+
+		Collection<Configuration> plus = providedScope.get("plus");
+
+		if (plus == null) {
+			plus = new ArrayList<>();
+		}
+
+		plus.add(targetPlatformIDEConfiguration);
+
+		providedScope.put("plus", plus);
+
+		scopes.put(GeneratedIdeaScope.PROVIDED.name(), providedScope);
+
+		ideaModule.setScopes(scopes);
+
+		SourceSet mainSourceSet = null;
+
+		try {
+			mainSourceSet = GradleUtil.getSourceSet(
+					project, SourceSet.MAIN_SOURCE_SET_NAME);
+		}
+		catch (UnknownDomainObjectException udoe) {
+			mainSourceSet = GradleUtil.addSourceSet(
+					project, SourceSet.MAIN_SOURCE_SET_NAME);
+		}
+
+		FileCollection compileClasspath = mainSourceSet.getCompileClasspath();
+
+		compileClasspath.plus(targetPlatformIDEConfiguration);
+	}
 }